+++ tests/phpunit/WikispeechHooksTest.php
@@ -15,8 +15,14 @@ class WikispeechHooksTest extends MediaWikiTestCase {
 	protected function setUp() : void {
 		parent::setUp();
 		$this->setMwGlobals( [
-				'wgWikispeechSpeechoidUrl' => 'https://server.domain',
-				'wgWikispeechVoices' => [ 'en' => 'en-voice' ]
+			'wgWikispeechSpeechoidUrl' => 'https://server.domain',
+			'wgWikispeechVoices' => [ 'en' => 'en-voice' ],
+			'wgWikispeechNamespaces' => [ NS_MAIN ],
+			'wgWikispeechKeyboardShortcuts' => 'shortcuts',
+			'wgWikispeechContentSelector' => 'selector',
+			'wgWikispeechSkipBackRewindsThreshold' => 'threshold',
+			'wgWikispeechHelpPage' => 'help',
+			'wgWikispeechFeedbackPage' => 'feedback'
 		] );
 		$context = new RequestContext();
 		$context->setLanguage( 'en' );
@@ -25,31 +31,49 @@ class WikispeechHooksTest extends MediaWikiTestCase {
 		$this->out->setTitle( $title );
 		$this->out->setRevisionId( $title->getLatestRevId() );
 		$this->out->getUser()->setOption( 'wikispeechEnable', true );
+		$this->out->getUser()->setOption( 'wikispeechShowPlayer', true );
 		MediaWikiServices::getInstance()
 			->getPermissionManager()
 			->overrideUserRightsForTesting(
 				$this->out->getUser(),
 				'wikispeech-listen'
 			);
-		$skinFactory = MediaWikiServices::getInstance()->getSkinFactory();
-		$this->skin = $skinFactory->makeSkin( 'fallback' );
+		$this->skin = $this->createStub( SkinTemplate::class );
+		$this->skin->method( 'getOutput' )->willReturn( $this->out );
 	}
 
 	public function testOnBeforePageDisplayLoadModules() {
 		Hooks::run( 'BeforePageDisplay', [ &$this->out, $this->skin ] );
 		$this->assertContains( 'ext.wikispeech', $this->out->getModules() );
+		$this->assertTrue( $this->configLoaded() );
+	}
+
+	private function configLoaded() {
+		$config = $this->out->getJsConfigVars();
+		return isset( $config['wgWikispeechKeyboardShortcuts'] ) &&
+			$config['wgWikispeechKeyboardShortcuts'] == 'shortcuts' &&
+			isset( $config['wgWikispeechContentSelector'] ) &&
+			$config['wgWikispeechContentSelector'] == 'selector' &&
+			isset( $config['wgWikispeechSkipBackRewindsThreshold'] ) &&
+			$config['wgWikispeechSkipBackRewindsThreshold'] == 'threshold' &&
+			isset( $config['wgWikispeechHelpPage'] ) &&
+			$config['wgWikispeechHelpPage'] == 'help' &&
+			isset( $config['wgWikispeechFeedbackPage'] ) &&
+			$config['wgWikispeechFeedbackPage'] == 'feedback';
 	}
 
 	public function testOnBeforePageDisplayDontLoadModulesIfWrongNamespace() {
 		$this->out->setTitle( Title::newFromText( 'Page', NS_TALK ) );
 		Hooks::run( 'BeforePageDisplay', [ &$this->out, $this->skin ] );
 		$this->assertEmpty( $this->out->getModules() );
+		$this->assertFalse( $this->configLoaded() );
 	}
 
-	public function testOnBeforePageDisplayDontLoadModulesIfDisabled() {
+	public function testOnBeforePageDisplayDontLoadModulesIfWikispeechDisabled() {
 		$this->out->getUser()->setOption( 'wikispeechEnable', false );
 		Hooks::run( 'BeforePageDisplay', [ &$this->out, $this->skin ] );
 		$this->assertEmpty( $this->out->getModules() );
+		$this->assertFalse( $this->configLoaded() );
 	}
 
 	public function testOnBeforePageDisplayDontLoadModulesIfLackingRights() {
@@ -58,6 +82,7 @@ class WikispeechHooksTest extends MediaWikiTestCase {
 			->overrideUserRightsForTesting( $this->out->getUser(), [] );
 		Hooks::run( 'BeforePageDisplay', [ &$this->out, $this->skin ] );
 		$this->assertEmpty( $this->out->getModules() );
+		$this->assertFalse( $this->configLoaded() );
 	}
 
 	public function testOnBeforePageDisplayDontLoadModulesIfServerUrlInvalid() {
@@ -67,6 +92,7 @@ class WikispeechHooksTest extends MediaWikiTestCase {
 		);
 		Hooks::run( 'BeforePageDisplay', [ &$this->out, $this->skin ] );
 		$this->assertEmpty( $this->out->getModules() );
+		$this->assertFalse( $this->configLoaded() );
 	}
 
 	public function testOnBeforePageDisplayDontLoadModulesIfRevisionNotAccessible() {
@@ -74,11 +100,83 @@ class WikispeechHooksTest extends MediaWikiTestCase {
 		$this->out->setRevisionId( $inaccessibleRevisionId );
 		Hooks::run( 'BeforePageDisplay', [ &$this->out, $this->skin ] );
 		$this->assertEmpty( $this->out->getModules() );
+		$this->assertFalse( $this->configLoaded() );
 	}
 
 	public function testOnBeforePageDisplayDontLoadModulesIfInvalidPageLanguage() {
 		$this->out->getContext()->setLanguage( 'sv' );
 		Hooks::run( 'BeforePageDisplay', [ &$this->out, $this->skin ] );
 		$this->assertEmpty( $this->out->getModules() );
+		$this->assertFalse( $this->configLoaded() );
+	}
+
+	public function testOnBeforePageDisplay_showPlayerNotSet_loadLoader() {
+		$this->out->getUser()->setOption( 'wikispeechShowPlayer', false );
+		Hooks::run( 'BeforePageDisplay', [ &$this->out, $this->skin ] );
+		$this->assertContains(
+			'ext.wikispeech.loader',
+			$this->out->getModules()
+		);
+		$this->assertNotContains( 'ext.wikispeech', $this->out->getModules() );
+		$this->assertTrue( $this->configLoaded() );
+	}
+
+	public function testOnSkinTemplateNavigation_addListenTab() {
+		// This stubbing is required to not get an error about Message::text().
+		$this->skin->method( 'msg' )->willReturn(
+			Message::newFromKey( 'wikispeech-listen' )
+		);
+		$this->out->getUser()->setOption( 'wikispeechShowPlayer', false );
+		$links = [ 'actions' => [] ];
+		Hooks::run( 'SkinTemplateNavigation', [ $this->skin, &$links ] );
+		$this->assertArrayHasKey( 'listen', $links['actions'] );
+	}
+
+	public function testOnSkinTemplateNavigation_wikispeechDisabled_dontAddListenTab() {
+		$this->out->getUser()->setOption( 'wikispeechEnable', false );
+		$links = [ 'actions' => [] ];
+		Hooks::run( 'SkinTemplateNavigation', [ $this->skin, &$links ] );
+		$this->assertArrayNotHasKey( 'listen', $links['actions'] );
+	}
+
+	public function testOnSkinTemplateNavigation_lackingRights_dontAddListenTab() {
+		MediaWikiServices::getInstance()
+			->getPermissionManager()
+			->overrideUserRightsForTesting( $this->out->getUser(), [] );
+		$links = [ 'actions' => [] ];
+		Hooks::run( 'SkinTemplateNavigation', [ $this->skin, &$links ] );
+		$this->assertArrayNotHasKey( 'listen', $links['actions'] );
+	}
+
+	public function testOnSkinTemplateNavigation_serverUrlInvalid_dontAddListenTab() {
+		$this->setMwGlobals(
+			'wgWikispeechSpeechoidUrl',
+			'invalid-url'
+		);
+		$links = [ 'actions' => [] ];
+		Hooks::run( 'SkinTemplateNavigation', [ $this->skin, &$links ] );
+		$this->assertArrayNotHasKey( 'listen', $links['actions'] );
+	}
+
+	public function testOnSkinTemplateNavigation_wrongNamespace_dontAddListenTab() {
+		$this->out->setTitle( Title::newFromText( 'Page', NS_TALK ) );
+		$links = [ 'actions' => [] ];
+		Hooks::run( 'SkinTemplateNavigation', [ $this->skin, &$links ] );
+		$this->assertArrayNotHasKey( 'listen', $links['actions'] );
+	}
+
+	public function testOnSkinTemplateNavigation_revisionNotAccessible_dontAddListenTab() {
+		$inaccessibleRevisionId = $this->out->getTitle()->getLatestRevId() - 1;
+		$this->out->setRevisionId( $inaccessibleRevisionId );
+		$links = [ 'actions' => [] ];
+		Hooks::run( 'SkinTemplateNavigation', [ $this->skin, &$links ] );
+		$this->assertArrayNotHasKey( 'listen', $links['actions'] );
+	}
+
+	public function testOnSkinTemplateNavigation_invalidPageLanguage_dontAddListenTab() {
+		$this->out->getContext()->setLanguage( 'sv' );
+		$links = [ 'actions' => [] ];
+		Hooks::run( 'SkinTemplateNavigation', [ $this->skin, &$links ] );
+		$this->assertArrayNotHasKey( 'listen', $links['actions'] );
 	}
 }